move registration header down one level
