+++ remote_config/src/include/firebase/remote_config/config_update_listener_registration.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef FIREBASE_REMOTE_CONFIG_SRC_INCLUDE_FIREBASE_CONFIG_UPDATE_LISTENER_REGISTRATION_H_
-#define FIREBASE_REMOTE_CONFIG_SRC_INCLUDE_FIREBASE_CONFIG_UPDATE_LISTENER_REGISTRATION_H_
+#ifndef FIREBASE_REMOTE_CONFIG_SRC_INCLUDE_FIREBASE_REMOTE_CONFIG_CONFIG_UPDATE_LISTENER_REGISTRATION_H_
+#define FIREBASE_REMOTE_CONFIG_SRC_INCLUDE_FIREBASE_REMOTE_CONFIG_CONFIG_UPDATE_LISTENER_REGISTRATION_H_
 
 #include <functional>
 
@@ -44,4 +44,4 @@ class ConfigUpdateListenerRegistration {
 }  // namespace remote_config
 }  // namespace firebase
 
-#endif  // FIREBASE_REMOTE_CONFIG_SRC_INCLUDE_FIREBASE_CONFIG_UPDATE_LISTENER_REGISTRATION_H_
+#endif  // FIREBASE_REMOTE_CONFIG_SRC_INCLUDE_FIREBASE_REMOTE_CONFIG_CONFIG_UPDATE_LISTENER_REGISTRATION_H_
