+++ remote_config/src/ios/remote_config_ios.h
@@ -20,6 +20,7 @@
 #include "app/src/reference_counted_future_impl.h"
 #include "app/src/util_ios.h"
 #include "firebase/future.h"
+
 #include "remote_config/src/include/firebase/remote_config.h"
 
 #ifdef __OBJC__
